+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 -   added `bg()` for add image/color backgrounds.
 -   added `start()` for automatically load the novel scene.
 -   added `VisualAction` with the methods:
-    - `fadeIn()` for fade when it appears.
+    -   `fadeIn()` for fade when it appears.
+-   added `AudioAction` with the methods:
+    -   `withVolume(volume: number)`
 
 ## Technical and misc